chore: design updates
