+++ packages/entities/entities-plugins/src/locales/en.json
@@ -1189,11 +1189,16 @@
       "realm_group_label": "Realms (can only select one)"
     },
     "kong_identity": {
-      "title": "Who should manage authentication and issue API keys?",
+      "title": "Who should manage authentication and credentials?",
+      "description_with_realms": "Use Consumers or centrally managed consumers for Consumer-based authentication. Use Kong Identity for principal-based authentication and advanced integrations.",
+      "description_without_realms": "Use Consumers for Consumer-based authentication. Use Kong Identity for principal-based authentication and advanced integrations.",
+      "learn_more": "Learn more.",
       "kong_identity_label": "Kong Identity",
-      "kong_identity_description": "Use Kong to generate, authenticate, and authorize API access. Kong Identity issues and manages API keys.",
+      "kong_identity_description_key_auth": "Use Kong Identity to authenticate API keys and map requests to a Kong Identity principal. The authenticated principal is added to the request context for advanced integrations.",
+      "kong_identity_description_basic_auth": "Use Kong Identity to authenticate Basic Auth credentials and map requests to a Kong Identity principal. The authenticated principal is added to the request context for advanced integrations.",
       "consumers_label": "Consumers",
-      "consumers_description": "Use consumer-managed API keys. Consumers provide their API key in a request header, query parameter, or request body to authenticate requests.",
+      "consumers_description_key_auth": "Use consumer-managed API keys. Consumers provide their API key in a request header, query parameter, or request body to authenticate requests.",
+      "consumers_description_basic_auth": "Use consumer-managed Basic Auth credentials. Consumers authenticate requests using a username and password.",
       "centrally_managed_label": "Centrally managed consumers",
       "centrally_managed_description": "Authenticate centrally managed consumers in Konnect using identity realms. Identity realms are scoped to the control plane by default."
     }
